+++ Source/Grimmetropolis/Grimmetropolis/Scripts/Game/ResourcePile.cs
@@ -1,15 +1,15 @@
 ﻿public class ResourcePile
 {
-    public float Wood { get; set; }
-    public float Stone { get; set; }
+    public int Wood { get; set; }
+    public int Stone { get; set; }
 
-    public ResourcePile(float wood, float stone)
+    public ResourcePile(int wood, int stone)
     {
         Wood = wood;
         Stone = stone;
     }
 
-    public ResourcePile() : this(0f, 0f) { }
+    public ResourcePile() : this(0, 0) { }
 
     public static ResourcePile operator -(ResourcePile a) => new ResourcePile(-a.Wood, -a.Stone);
     public static ResourcePile operator +(ResourcePile a, ResourcePile b) => new ResourcePile(a.Wood + b.Wood, a.Stone + b.Stone);
